many: move chooser to initramfs
It was decided in London that chooser should be launched from initramfs,
but also be available in the normal system as a means to set up the next
boot of the system.

This change moves the chooser to initramfs, but does not implement the
running system setup capabilities yet, also because the exact mechanism
of setting up the next boot is still undefined.

In initramfs chooser is simpler and relies on reading recovery versions
and writing the chosen recovery version from and to the filesystem
(which removes many ugly hacks from the original implementation).

Note: also requires changes in core-build to actually call the new
chooser from initramfs.

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>